app/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=app/requirements.txt pyproject.toml
#
apispec[marshmallow,yaml]==6.6.1
    # via
    #   apispec-webframeworks
    #   flask-apispec
    #   matcha-back (pyproject.toml)
apispec-webframeworks==1.1.0
    # via matcha-back (pyproject.toml)
authlib==1.3.1
    # via matcha-back (pyproject.toml)
blinker==1.8.2
    # via flask
certifi==2024.7.4
    # via requests
cffi==1.16.0
    # via cryptography
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via flask
cryptography==42.0.8
    # via authlib
flask==3.0.3
    # via
    #   flask-apispec
    #   flask-cors
    #   flask-jwt-extended
    #   flask-swagger-ui
    #   matcha-back (pyproject.toml)
flask-apispec==0.11.4
    # via matcha-back (pyproject.toml)
flask-cors==4.0.1
    # via matcha-back (pyproject.toml)
flask-jwt-extended==4.6.0
    # via matcha-back (pyproject.toml)
flask-swagger-ui==4.11.1
    # via matcha-back (pyproject.toml)
idna==3.7
    # via requests
iniconfig==2.0.0
    # via pytest
itsdangerous==2.2.0
    # via flask
jinja2==3.1.4
    # via flask
markupsafe==2.1.5
    # via
    #   jinja2
    #   werkzeug
marshmallow==3.21.3
    # via
    #   apispec
    #   flask-apispec
    #   matcha-back (pyproject.toml)
    #   webargs
packaging==24.1
    # via
    #   apispec
    #   marshmallow
    #   pytest
    #   webargs
pluggy==1.5.0
    # via pytest
psycopg2-binary==2.9.9
    # via matcha-back (pyproject.toml)
pycparser==2.22
    # via cffi
pyjwt==2.8.0
    # via flask-jwt-extended
pytest==8.2.2
    # via matcha-back (pyproject.toml)
python-dotenv==1.0.1
    # via matcha-back (pyproject.toml)
pyyaml==6.0.1
    # via apispec
requests==2.32.3
    # via matcha-back (pyproject.toml)
sqlalchemy==2.0.31
    # via matcha-back (pyproject.toml)
typing-extensions==4.12.2
    # via sqlalchemy
urllib3==2.2.2
    # via requests
uwsgi==2.0.26
    # via matcha-back (pyproject.toml)
webargs==8.4.0
    # via flask-apispec
werkzeug==3.0.3
    # via
    #   flask
    #   flask-jwt-extended
